Moved consts.rs to proto/src/.
