+++ components/funder/src/token_channel.rs
@@ -8,7 +8,7 @@ use crypto::hash::sha_512_256;
 use identity::IdentityClient;
 
 use proto::funder::messages::{MoveToken, FriendTcOp, MoveTokenRequest};
-use proto::funder::consts::MAX_OPERATIONS_IN_BATCH;
+use proto::consts::MAX_OPERATIONS_IN_BATCH;
 use proto::funder::signature_buff::verify_friend_move_token;
 
 use crate::mutual_credit::types::{MutualCredit, McMutation};
