+++ components/proto/src/funder/mod.rs
@@ -1,4 +1,3 @@
 pub mod messages;
-pub mod consts;
 pub mod signature_buff;
 pub mod serialize;
